+++ src/ocd/date.rs
@@ -9,10 +9,9 @@ use std::path::PathBuf;
 use std::sync::LazyLock;
 
 // The default date regex string.
-pub const DATE_FLORB_REGEX_STR: &str = r"(?<date>[0-9];{4}.?[0-9]{2}.?[0-9]{2}|(?:(?:\d{1,2})\s(?i)(?:jan|january|feb|february|mar|march|apr|april|may|jun|june|jul|july|aug|august|sep|september|oct|october|nov|november|dec|december)\s(?:\d{1,4})))";
-pub static DATE_FLORB_REGEX: LazyLock<Regex> =
+pub(crate) const DATE_FLORB_REGEX_STR: &str = r"(?<date>[0-9];{4}.?[0-9]{2}.?[0-9]{2}|(?:(?:\d{1,2})\s(?i)(?:jan|january|feb|february|mar|march|apr|april|may|jun|june|jul|july|aug|august|sep|september|oct|october|nov|november|dec|december)\s(?:\d{1,4})))";
+pub(crate) static DATE_FLORB_REGEX: LazyLock<Regex> =
     LazyLock::new(|| Regex::new(DEFAULT_DATEFINDER_REGEX_STR).unwrap());
-// pub const DEFAULT_DATE_REGEX_STR: &str = r"(?<a>(?<y1>1\d\d\d|20\d\d).?(?<m1>0[1-9]|1[012]).?(?<d1>0[1-9]|[12]\d|30|31))|(?i)(?<b>(?<d2>\d{1,2})\s(?<m2>jan|january|feb|february|mar|march|apr|april|may|jun|june|jul|july|aug|august|sep|september|oct|october|nov|november|dec|december)\s(?<y2>\d{1,4}))";
 
 /// The default datefinder reges string is the same as the default date regex but includes non-alphanumeric catch-all patterns before and after.
 /// Case A: a date in the format YYYY?MM?DD or YYYYMMDD
@@ -20,24 +19,18 @@ pub static DATE_FLORB_REGEX: LazyLock<Regex> =
 /// Case B: case insensitive, DD MONTH YYYY
 /// where MONTH may be the full month name or the three letter short version.
 /// `(?i)(?<b>(?<d2>\d{1,2})\s(?<m2>jan|january|feb|february|mar|march|apr|april|may|jun|june|jul|july|aug|august|sep|september|oct|october|nov|november|dec|december)\s(?<y2>\d{1,4}))`
-pub const DEFAULT_DATEFINDER_REGEX_STR: &str = r"\D*(?<a>(?<y1>1\d\d\d|20\d\d).?(?<m1>0[1-9]|1[012]).?(?<d1>0[1-9]|[12]\d|30|31))|(?i)(?<b>(?<d2>\d{1,2})\s(?<m2>jan|january|feb|february|mar|march|apr|april|may|jun|june|jul|july|aug|august|sep|september|oct|october|nov|november|dec|december)\s(?<y2>\d{1,4}))\D*";
-pub static DEFAULT_DATEFINDER_REGEX: LazyLock<Regex> =
+const DEFAULT_DATEFINDER_REGEX_STR: &str = r"\D*(?<a>(?<y1>1\d\d\d|20\d\d).?(?<m1>0[1-9]|1[012]).?(?<d1>0[1-9]|[12]\d|30|31))|(?i)(?<b>(?<d2>\d{1,2})\s(?<m2>jan|january|feb|february|mar|march|apr|april|may|jun|june|jul|july|aug|august|sep|september|oct|october|nov|november|dec|december)\s(?<y2>\d{1,4}))\D*";
+pub(crate) static DEFAULT_DATEFINDER_REGEX: LazyLock<Regex> =
     LazyLock::new(|| Regex::new(DEFAULT_DATEFINDER_REGEX_STR).unwrap());
 
-#[derive(Debug, Default)]
-pub enum DateFlorbRegex {
-    #[default]
-    Default,
-}
-
 #[derive(Debug, PartialEq)]
-pub enum DateSource {
+pub(crate) enum DateSource {
     Filename,
     Exif,
     Filesystem,
 }
 
-pub fn regex_date(haystack: &str) -> Option<(u32, u32, u32)> {
+pub(crate) fn regex_date(haystack: &str) -> Option<(u32, u32, u32)> {
     DEFAULT_DATEFINDER_REGEX.captures(haystack).map(|capture| {
         if capture.name("a").is_some() {
             let year = capture.get(2).unwrap().as_str().parse::<u32>().unwrap();
@@ -81,7 +74,7 @@ fn english_month_to_number(month: &str) -> u32 {
 }
 
 /// Given a filename, extracts a date by matching against a regex.
-pub fn filename_date(file_name: &Path) -> Option<(DateSource, u32, u32, u32)> {
+pub(crate) fn filename_date(file_name: &Path) -> Option<(DateSource, u32, u32, u32)> {
     file_name
         .to_str()
         .and_then(crate::ocd::date::regex_date)
@@ -98,7 +91,7 @@ pub fn filename_date(file_name: &Path) -> Option<(DateSource, u32, u32, u32)> {
 /// - if for some reason the exif tag is not in the right format and
 ///   `chrono::NaiveDateTime::parse_from_str` cannot parse it, parse the result
 ///   with `dateparser::parse` as a date with format `"%Y-%m-%d`.
-pub fn exif_date(path: &PathBuf) -> Option<(DateSource, u32, u32, u32)> {
+pub(crate) fn exif_date(path: &PathBuf) -> Option<(DateSource, u32, u32, u32)> {
     std::fs::File::open(path).ok().and_then(|file| {
         let mut bufreader = std::io::BufReader::new(&file);
         exif::Reader::new()
@@ -138,7 +131,7 @@ pub fn exif_date(path: &PathBuf) -> Option<(DateSource, u32, u32, u32)> {
 /// - obtain the file metadata
 /// - get the `created` field
 /// - check whether the created is the same as the current date
-pub fn metadata_date(path: &PathBuf) -> Option<(DateSource, u32, u32, u32)> {
+pub(crate) fn metadata_date(path: &PathBuf) -> Option<(DateSource, u32, u32, u32)> {
     std::fs::metadata(path).ok().and_then(|metadata| {
         metadata.created().ok().and_then(|system_time| {
             let today: chrono::DateTime<chrono::offset::Local> = chrono::Local::now();
@@ -154,4 +147,34 @@ pub fn metadata_date(path: &PathBuf) -> Option<(DateSource, u32, u32, u32)> {
             }
         })
     })
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use std::path::Path;
+
+    #[test]
+    fn filename_date1() {
+        let file_name = Path::new("An image file from 2024-12-31.jpg");
+        let expected = Some((DateSource::Filename, 2024, 12, 31));
+        let result = filename_date(file_name);
+        assert_eq!(expected, result);
+    }
+
+    #[test]
+    fn filename_date2() {
+        let file_name = Path::new("An image file from 20241231.jpg");
+        let expected = Some((DateSource::Filename, 2024, 12, 31));
+        let result = filename_date(file_name);
+        assert_eq!(expected, result);
+    }
+
+    #[test]
+    fn filename_date3() {
+        let file_name = Path::new("An image file from 2024-12-01 to 2024-12-31.jpg");
+        let expected = Some((DateSource::Filename, 2024, 12, 1));
+        let result = filename_date(file_name);
+        assert_eq!(expected, result);
+    }
 }
